+++ CHANGELOG.md
@@ -9,6 +9,14 @@
 
 - Added support for Android 34
 
+## Fixes
+
+### Platform Specific Fixes
+
+#### iOS
+
+- Fixed Call Messages not being built with the passed `contentType` or `messageType`.
+
 ## Changes
 
 ### Call Message Events
@@ -28,6 +36,12 @@ call.sendMessage(new CallMessage({
   messageType: 'user-defined-message',
 }));
 ```
+
+- Added new error codes. See the following table for details:
+  | Error Code | Description |
+  | --- | --- |
+  | 31210 | Raised when a Call Message is sent with an invalid message type. |
+  | 31211 | Raised when a Call Message is sent when the call is not yet ready to send messages. This can typically happen when the Call/CallInvite is not yet in a ringing state. |
 
 1.0.0 (Mar 25, 2024)
 ====================